\name{bmn.ir}
\alias{bmn.ir}
\docType{data}
\title{
Iran Broad Mony data.
}
\description{
Iran Broad Mony data. Broad money (current LCU)
}
\usage{bmn.ir}
\format{
  The format is:
 chr "bmn.ir"
 \itemize{
  \item{year}{year}
  \item{bmn} {Broad money}
 }
}
\details{
Iran Broad Mony data. Broad money (current LCU) from 1960 to 2010.
}
\source{http://www.worldbank.org/
}
\references{worldbank.com
}
\examples{
data(bmn.ir)
## maybe str(bmn.ir) ; plot(bmn.ir) ...
}
\keyword{datasets}
