\name{cow}
\alias{cow}
\docType{data}
\title{Cow Data
}
\description{
Milk production amount for a single cow within a year.
}
\usage{data("cow")}
\format{
  The format is: data.frame
 chr "cow"
 \itemize{
    \item Day: Day from 1 to end of the year.
    \item Milk: Milk produced in Litre.
 }

}
\details{
The data collected within a year of a milk roductin from a fluk of cows. \code{cow} variable is reported for a single cow.
}
\source{
A cow farm production in Fars province of iran.
}
\references{
   Riazoshams H, Midi H, and Ghilagaber G, 2018,. Robust Nonlinear Regression, with Application using R, Joh Wiley and Sons.
}
\examples{
data(cow)
}
\keyword{datasets}
