\name{DrugKenakin}
\alias{DrugKenakin}
\docType{data}
\title{
Responses to the concentration of an agonist in a functional assay. 
}
\description{
Kenakin used a set of responses to the concentration of an agonist in a functional assay. 
}
\usage{data("DrugKenakin")}
\format{
  The format is: data.frame
 chr "DrugKenakin"
   \itemize{
    \item {Concentration} {Concentration, predictor.}
    \item {Response} {response, response variable.}
}
}
\details{
Kenakin used a set of responses to the concentration of an agonist in a functional assay. They fit the following model to their data. In this data, observation 5 has an outlier in the response direction.
}
\source{
Kenakin TP. A Pharmacology Primer: Theory, Applications, and Methods. Third Edition Academic Press; 2009. pp. 286-287.
}
\references{
Kenakin TP. A Pharmacology Primer: Theory, Applications, and Methods. Third Edition Academic Press; 2009. pp. 286-287.
}
\examples{
data(DrugKenakin)
DrugKenakin
}
\keyword{datasets}