\name{I}
\alias{I}

\title{
 Inhibit Interpretation/conversion of object
}
\description{
 This is \code{AsIs} function for compatibility purpose. See \code{AsIs} help.
 }
\usage{
  I(x)
}
\arguments{
  \item{x}{
    an object
  }
}
\details{
Function has two main usage.
  See \code{AsIs} help.
}
\value{
  A copy of object with class "AsIs" prepared to the class(es).
}
\references{
  Chambers, J. M. (1992) Linear models. Chapter 4 of statistical models in S eds J. M. Chamber and T. J. Hastie, Wadsworth & Brooks/Cole.
}
\author{
  Chambers, J. M. (1992)
}
\note{
  This is I, and AsIs function in base package, inserted here for compatibility to change of package versions.
}

\seealso{
\code{AsIs}
}
\examples{
  ## The function is currently defined as
  "I"
  }
\keyword{I}

