\name{net.ir}
\alias{net.ir}
\docType{data}
\title{Iran Net Money Data.
}
\description{Net domestic credit (current LCU) Iran.
}
\usage{net.ir}
\format{
  The format is: data.frame
  chr "net.ir"
  \itemize{
    \item {year:} {year}
    \item {net:} {net money amount.}
 }
}
\details{Net domestic credit (current LCU) of Iran, from 1961.
}
\source{http://www.worldbank.org/
}
\references{worldbank.com
}
\examples{
data(net.ir)       # load data
net.ir             # access by variable name
}
\keyword{datasets}
