\name{net.sw}
\alias{net.sw}
\docType{data}
\title{Sweden Net Money Data.
}
\description{Net domestic credit (current LCU) Sweden.
}
\usage{net.sw}
\format{
  The format is: data.frame
 chr "net.sw"
   \itemize{
    \item year: year
    \item net: net money amount.
 }
}
\details{Net domestic credit (current LCU) of Sweden.
}
\source{ \url{http://www.worldbank.org/}
obtained 
}
\references{worldbank.com
}
\examples{
data(net.sw)
net.sw
}
\keyword{datasets}
