\name{nl.fittorNULL}
\Rdversion{1.1}
\docType{class}
\alias{nl.fittorNULL}
\alias{functionorNULL}
\alias{expressionorNULL}
\alias{callorNULL}
\alias{integerorNULL}
\alias{numericorNULL}
\alias{characterorNULL}
\alias{logicalorNULL}
\alias{listorNULL}
\alias{nl.fitt.roborNULL}
\alias{fittmethodorNULL}
\alias{vectororNULL}
\alias{matrixororNULL}
\alias{vectororMatrix}


\title{"OR" Class}
\description{
  Set of "or" classes are created for compatibility reasons.
  \itemize{
  \item{nl.fittorNULL}{ \code{\link{nl.fitt}} object or \code{NULL}}
  \item{functionorNULL}{ \code{function} object or \code{NULL}}
  \item{expressionorNULL}{ \code{expression} object or \code{NULL}}
  \item{callorNULL}{\code{call} object or \code{NULL}}
  \item{integerorNULL}{\code{integer} object or \code{NULL}}
  \item{numericorNULL}{\code{numeric} object or \code{NULL}}
  \item{characterorNULL}{\code{character} object or \code{NULL}}
  \item{logicalorNULL}{\code{logical} object or \code{NULL}}
  \item{listorNULL}{\code{list} object or \code{NULL}}

  \item{nl.fitt.roborNULL}{\code{\link{nl.fitt.rob}} object or \code{NULL}}
  \item{fittmethodorNULL}{\code{\link{fittmethod}} object or \code{NULL}}
  \item{vectororNULL}{\code{vector} object or \code{NULL}}
  \item{matrixororNULL}{\code{matrix} object or \code{NULL}}
  \item{vectororMatrix}{\code{vector} object or \code{matrix}}
}
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "nl.fittorNULL" in the signature.
}
\references{
   Riazoshams H, Midi H, and Ghilagaber G, 2018,. Robust Nonlinear Regression, with 
Application using R, Joh Wiley and Sons.
}
\author{
  Hossein Riazoshams, May 2014. 
  Email: \email{riazihosein@gmail.com} 
  URL \url{http://www.riazoshams.com/nlr/}
}
\note{
  These classes used in several slots of objects. User might not use them directly.
}
\seealso{
\code{\link{nl.fitt}}, \code{\link{nl.form}}, \code{\link{nl.fitt.rob}}, \code{\link{fittmethod}}
}
\examples{
showClass("nl.fittorNULL")
}
\keyword{classes}
