\name{nl.formorNULL}
\Rdversion{1.1}
\docType{class}
\alias{nl.formorNULL}

\title{Class \code{"nl.formorNULL"}}
\description{nl.formorNULL, Union class that can have \code{nl.form} value or NULL value.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{Union class used in assignment of slots, in nl.fitt.gn, and others.
No methods defined with class "nl.formorNULL" in the signature.
}
\references{
  Robust Nonlinear Regression, Theories and Methods with Practical Guides for R Packages. Riazoshams et al. 
}
\author{
  Hossein Riazoshams, Apr 2014. 
  Email: \email{riazihosein@gmail.com} 
  URL \url{www.riazoshams.com/nlr}
}

\note{
  For internal usage, it might not call directly by user.
}
\seealso{
  \code{\link{nl.form}}
}
\examples{
showClass("nl.formorNULL")
}
\keyword{classes}
