\name{tadr.ir}
\alias{tadr.ir}
\docType{data}
\title{
Iran Birth Rate.
}
\description{
Birth rate, crude (per 1,000 people).
}
\usage{tadr.ir}
\format{
  The format is:
 chr "tadr.ir"
    \itemize{
    \item {year:} {year}
    \item {tadr:} {birth rate per}
 }
}
\details{
Birth rate, crude (per 1,000 people). From 1960 to 2011.
}
\source{http://www.worldbank.org/
}
\references{worldbank.com
}
\examples{
data(tadr.ir)
tadr.ir
}
\keyword{datasets}
