\name{theophil}
\alias{theophil}
\docType{data}
\title{
  Iran Trade Data.
}
\description{
  Trademark applications, direct resident.
}
\usage{theophil}
\format{
  The format is:
    chr "theophil"
  \itemize{
    \item {time:} {time}
    \item {concenteration:} {theophil concenteration}
  }
}
\details{
  Theophil concentration, the two consecutive first order reaction fitts to data, 
  for which the model stored in \code{\link{fktlistex6}} object.
}
\source{
  Unknown web, 2007.
}
\references{
   Riazoshams H, Midi H, and Ghilagaber G, 2018,. Robust Nonlinear Regression, with 
Application using R, Joh Wiley and Sons.
}
\examples{
  data(theophil)
  theophil
}
\keyword{datasets}
