\name{trade.ir}
\alias{trade.ir}
\docType{data}
\title{
Iran Trade Data.
}
\description{
Trademark applications, direct resident.
}
\usage{trade.ir}
\format{
  The format is:
 chr "trade.ir"
   \itemize{
    \item {year:} {year}
    \item {trade:} {trademark}
 }
}
\details{
Trademark applications, direct resident from 1960 to 2006.
}
\source{
http://www.worldbank.org/
}
\references{
worldbank.com
}
\examples{
data(trade.ir)
trade.ir
}
\keyword{datasets}
