\name{var1}
\alias{var1}
\title{
Compute Variance.
}
\description{
Compute variance of a vector.
}
\usage{
var1(x)
}
\arguments{
  \item{x}{
      Vector of data.
}
}
\details{
Only variance, for internal use.
}
\value{
Numeric Variance of a vector.}
\references{
  Robust Nonlinear Regression, Theories and Methods with Practical Guides for R Packages. Riazoshams et al. 
}
\author{
  Hossein Riazoshams, 2013. 
  Email: \email{riazihosein@gmail.com} 
  URL \url{www.riazoshams.com/nlr}
}
\note{
For internal use. might not be directly called by user.
}
\seealso{
\code{\link{nlr}}
}
\examples{
## The function is currently defined as
"var1"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Variance}
